requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.5.0
pydantic==2.4.2
openai==0.28.1
python-dotenv==1.0.0
bson==0.5.10
python-multipart==0.0.6
email-validator==2.0.0
websockets==12.0
aiofiles==23.2.1